requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
moviepy
psycopg2
webvtt-py
youtube_dl
